requirements.txt
chromadb>=1.5.4
langchain>=0.3.25
langgraph>=1.0.1
langchain-openai>=0.3.23
langchain-neo4j>=0.4.0
sentence-transformers==3.3.1
transformers>=4.57.0
huggingface_hub>=1.6.0
torch>=2.0.0
neo4j>=5.28.1
gradio>=6.6.0
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
